Refactoring start...again
