+++ lib/ui5Framework/npm/Installer.js
@@ -1,0 +1,65 @@
+const path = require("path");
+const fs = require("graceful-fs");
+const {promisify} = require("util");
+const stat = promisify(fs.stat);
+const readFile = promisify(fs.readFile);
+const Registry = require("./Registry");
+const log = require("@ui5/logger").getLogger("normalizer:ui5Framework:npm:Installer");
+
+class Installer {
+	constructor() {
+		const homedir = require("os").homedir();
+		this._baseDir = path.join(homedir, ".ui5", "framework", "packages");
+		this._cacheDir = path.join(homedir, ".ui5", "framework", "cacache");
+		log.verbose(`Installing to: ${this._baseDir}`);
+		this._registry = new Registry({cacheDir: this._cacheDir});
+
+		this._installedCounter = 0;
+		this._cachedCounter = 0;
+	}
+
+	async fetchPackageManifest({pkgName, version = this._version, targetDir = this._getTargetDirForPackage({pkgName, version})}) {
+		try {
+			const pkg = JSON.parse(await readFile(path.join(targetDir, "package.json"), {encoding: "utf8"}));
+			return {
+				name: pkg.name,
+				dependencies: pkg.dependencies,
+				devDependencies: pkg.devDependencies
+			};
+		} catch (err) {
+			if (err.code === "ENOENT") { // "File or directory does not exist"
+				const manifest = await this._registry.requestPackageManifest(pkgName, version);
+				return {
+					name: manifest.name,
+					dependencies: manifest.dependencies,
+					devDependencies: manifest.devDependencies
+				};
+			} else {
+				throw err;
+			}
+		}
+	}
+
+	async installPackage({pkgName, version, targetDir = this._getTargetDirForPackage({pkgName, version})}) {
+		try {
+			await stat(path.join(targetDir, "package.json"));
+			log.verbose(`Alrady installed: ${pkgName} in version ${version}`);
+			this._cachedCounter++;
+		} catch (err) {
+			if (err.code === "ENOENT") { // "File or directory does not exist"
+				log.info(`Installing ${pkgName}...`);
+				log.verbose(`Installing ${pkgName} in version ${version}...`);
+				await this._registry.extractPackage(pkgName, version, targetDir);
+				this._installedCounter++;
+			} else {
+				throw err;
+			}
+		}
+	}
+
+	_getTargetDirForPackage({pkgName, version}) {
+		return path.join(this._baseDir, ...pkgName.split("/"), version);
+	}
+}
+
+module.exports = Installer;
